requirements.txt
PyPDF2
nltk
sentence_transformers
faiss-cpu
numpy
requests
python-dotenv
flask
celery
redis
uuid
